remove empty folder and rename header folder  to be tabPanel Fix #23
