+++ src/component/tabPanel/style.js
